backend/requirements.txt
django
djangorestframework
django-rest-framework-nested
django-cors-headers
psycopg2-binary
drf-yasg
coreapi
Pillow
pyyaml
uwsgi
rich
